chore: move syslog to its own package
